Cascading task delete with JavaScript confirmation; small UI fixes
git-svn-id: http://agilefant.googlecode.com/svn/trunk@437 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ web/WEB-INF/jsp/editBacklogItem.jsp
@@ -240,7 +240,7 @@ example 3h) </td>
 					<ww:url id="deleteLink" action="deleteTask" includeParams="none">
 						<ww:param name="taskId" value="${row.id}"/>
 					</ww:url>
-					<ww:a href="%{editLink}">Edit</ww:a>|<ww:a href="%{deleteLink}">Delete</ww:a>
+					<ww:a href="%{editLink}">Edit</ww:a>|<ww:a href="%{deleteLink}" onclick="return confirmDeleteTask()">Delete</ww:a>
 				</display:column>
 			</display:table>
 		</p>
